# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tic Tac Toe API
  description: REST API for Tic Tac Toe game with WebSocket real-time updates
  version: 1.0.0
servers:
  - url: http://localhost:5150
    description: Development server
paths:
...
  /api/tile/{x}/{y}:
    post:
      summary: Make a move
      description: Attempt to place X or O at the specified coordinates
      parameters:
        - name: x
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 2
          description: X coordinate (0-2)
        - name: y
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 2
          description: Y coordinate (0-2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveRequest'
            example:
              state: "X"
      responses:
        '200':
          description: Move successful
        '409':
          description: Tile already occupied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Tile already occupied"
        '400':
          description: Invalid coordinates or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid coordinates"
...
components:
  schemas:
    TileState:
      type: string
      enum: ["Empty", "X", "O"]
      description: State of a tile in the game grid
...
    MoveRequest:
      type: object
      required:
        - state
      properties:
        state:
          $ref: '#/components/schemas/TileState'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message